Rename scanner to altbeacon for better reflection of the used tools
