+++ lib/mapper/altbeacon/beacon.js
@@ -18,7 +18,7 @@ function map(beaconRaw) {
         UUID: beaconRaw.uuid.toUpperCase(),
         major: parseInt(beaconRaw.major),
         minor: parseInt(beaconRaw.minor),
-        rssi: beaconRaw.rssi,
+        rssi: parseInt(beaconRaw.rssi, 10),
         accuracy: beaconRaw.accuracy,
         proximity: beaconRaw.proximity
     };
