+++ lib/mapper/altbeacon/beaconRegionMonitoring.js
@@ -5,7 +5,7 @@ var Alloy = require('alloy'),
 /* jshint ignore:end */
 
 /**
- * Create a beaconregion from a knownbeacon used by the drtechscanner
+ * Create a beaconregion from a knownbeacon used by the altbeaconscanner
  * @param knownBeacon A knownbeacon
  * @returns {{uuid: String, identifier: String}}
  */
