+++ lib/scanners/altbeacon.js
@@ -5,16 +5,16 @@ var Alloy = require('alloy'),
 /* jshint ignore:end */
 
 var BaseScanner = require('./../scanners/base');
-var beaconMapper = require('./../mapper/beuckman/beacon');
-var beaconRegionMapper = require('./../mapper/beuckman/beaconRegion');
-var beaconRegionMonitoringMapper = require('./../mapper/beuckman/beaconRegionMonitoring');
+var beaconMapper = require('./../mapper/altbeacon/beacon');
+var beaconRegionMapper = require('./../mapper/altbeacon/beaconRegion');
+var beaconRegionMonitoringMapper = require('./../mapper/altbeacon/beaconRegionMonitoring');
 
 /**
- * Drtech scanner to scan iBeacons on Android devices
+ * Altbeacon scanner to scan iBeacons on Android devices
  * @param boolean backgroundMode - Parameter to handle beacons when the application is running in backgroundmode
  * @returns {BaseScanner}
  */
-function drtechScanner(backgroundMode) {
+function altbeaconScanner(backgroundMode) {
     var self = new BaseScanner(beaconMapper, beaconRegionMapper, beaconRegionMonitoringMapper);
     self.Beacons = require('com.drtech.altbeacon');
 
@@ -67,4 +67,4 @@ function drtechScanner(backgroundMode) {
     return self;
 }
 
-module.exports = drtechScanner;
+module.exports = altbeaconScanner;
